requirements.txt
awscli==1.16.140
boto3==1.9.130
botocore==1.12.130
colorama==0.3.9
Django==2.2
django-elasticsearch-dsl==0.5.1
djangorestframework==3.9.2
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
ipaddress==1.0.22
jmespath==0.9.4
mysqlclient==1.4.2.post1
postgres==2.2.2
psycopg2-binary==2.8.2
pyasn1==0.4.5
python-dateutil==2.5.2
pytz==2018.9
PyYAML==3.11
rsa==3.4.2
ruamel.yaml==0.15.97
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
yml2json==1.1.3
